farewell to the BTC-e api
in the interest of posterity:

due to the shadyness of this exchange, it had low liquidity and
always lacked market makers. for the few weeks that experiments
were conducted for the purpose of testing the code deleted by
this commit, the per-volume-traded profitability of scalping
on this exchange far outshone all the other exchanges.

shady mobster exchange, you will be dearly missed!