slightly improve nomenclature (1/N)
the original implied that the effect would be linear; however, the
actual usage of this value is beyond linear. arguably the new name
isn't completely accurate either as the intuition which should be
targeted by the nomenclature is that of the final effect including
the distribution of taker flow...

even if the nomenclature remains, the initarg name must also be
updated eventually; that will require modifying the scripts and
is thus postponed for a few days.